requirements.txt
langchain==0.0.300
langchain-community
gitpython
fastapi
uvicorn
faiss-cpu
redis
celery
huggingface-hub
google-cloud-aiplatform
click
pytest